% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aberrant_values.R
\name{aberrant_values}
\alias{aberrant_values}
\title{Create a summary table which, for each row, displays the rows name, type, number of unique values, and number and percentage of missing values.}
\usage{
aberrant_values(df)
}
\arguments{
\item{df}{the \code{data.frame} to create a summary from}
}
\value{
A \code{tbl_df} with one row per column of \code{df} and columns (column, class, n_distinct, n_NA, n_not_NA, pct_NA)
}
\description{
Create a summary table which, for each row, displays the rows name, type, number of unique values, and number and percentage of missing values.
}

