% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/categorical_summary.R
\name{categorical_summary}
\alias{categorical_summary}
\title{Create a summary table of a data.frame's categorical (factor) variables.}
\usage{
categorical_summary(df)
}
\arguments{
\item{df}{The \code{data.frame} to create a categorical summary from}
}
\value{
A \code{tbl_df} with two columns, \code{variable} and \code{levels}. The \code{variable} column contains the name of each categorical column, repeated for each level it had, in the same order as in \code{df}. The \code{levels} column contains each level for the associated \code{variable}.
}
\description{
This will output a two-column table where the first is the name of the column and the second is the name of that column's levels.
}
\examples{
df <- data_frame(id = 1:5, gender = factor(c("M", "F", "M", "M", "F")), race = factor(c("white", "white", "black", "black", "other")))
categorical_summary(df)

}

