% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deidentify.R
\name{deidentify}
\alias{deidentify}
\title{Deidentify a single data frame.}
\usage{
deidentify(df, mask = NULL, drop = NULL)
}
\arguments{
\item{df}{(\code{data.frame}) the dataset to deidentify}

\item{mask}{(\code{character}) names of columns in \code{df} to mask}

\item{drop}{(\code{character}) names of columns in \code{df} to drop}
}
\value{
\code{data.frame} \code{df} with drop columns removed and mask columns replaced
}
\description{
This removes sensitive columns from a dataset and masks ID columns by mapping levels to integers.
}
\examples{
deidentify(mtcars, mask = c("mpg", "cyl"), drop = c("vs", "am"))

}

