% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table_1.R
\name{table_1}
\alias{table_1}
\title{Create a "table 1" from a data.frame, a common summary table used in publications.}
\usage{
table_1(df, dimension = NULL, overall = TRUE)
}
\arguments{
\item{df}{the \code{data.frame} to summarize}

\item{dimension}{the name of a column in \code{df} to use as the dimensions; used to form columns of the output}

\item{overall}{should an overall column be generated?}
}
\value{

}
\description{
This will not return a ready-to-publish table, but does most of the hard work.
It should be straightforward to fix up the table returned by this function.
}

