% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zelig_prediction_summary.R
\name{zelig_prediction_summary}
\alias{zelig_prediction_summary}
\title{Extract summary simulation stats from a Zelig model}
\usage{
zelig_prediction_summary(model)
}
\arguments{
\item{model}{a Zelig model; class should start with "Zelig"}
}
\value{
\code{tbl_df} with one row per simulation-condition pair, and columns \code{simulation}, \code{condition}, \code{lower}, \code{mean}, \code{upper}
}
\description{
Returns simulation name, condition name, and the following summary stats of their expected values (ev):
lower quantile (2.5\%), mean (NOT median), and upper quantile (97.5\%).
}

