% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zelig_prediction_summary.R
\name{zelig_prediction_summary}
\alias{zelig_prediction_summary}
\title{Extract and combine summary simulation stats from a Zelig models}
\usage{
zelig_prediction_summary(model)
}
\arguments{
\item{models}{}
}
\description{
Extract and combine summary simulation stats from a Zelig models
}

